requirements.txt
pip~=25.0.1
wheel~=0.45.1
setuptools~=75.8.2
certifi~=2025.1.31
urllib3~=2.3.0
idna~=3.10
requests~=2.32.3
python-dotenv~=1.0.1